+++ clients/vip-loan/examples/rest-api/MarketDataApi/queryVIPLoanFixedRateMarket.ts
@@ -1,0 +1,26 @@
+import { VIPLoan, VIP_LOAN_REST_API_PROD_URL } from '../../../src';
+
+const configurationRestAPI = {
+    apiKey: process.env.API_KEY ?? '',
+    apiSecret: process.env.API_SECRET ?? '',
+    basePath: process.env.BASE_PATH ?? VIP_LOAN_REST_API_PROD_URL,
+};
+const client = new VIPLoan({ configurationRestAPI });
+
+async function queryVIPLoanFixedRateMarket() {
+    try {
+        const response = await client.restAPI.queryVIPLoanFixedRateMarket({
+            loanCoin: 'loanCoin_example',
+        });
+
+        const rateLimits = response.rateLimits!;
+        console.log('queryVIPLoanFixedRateMarket() rate limits:', rateLimits);
+
+        const data = await response.data();
+        console.log('queryVIPLoanFixedRateMarket() response:', data);
+    } catch (error) {
+        console.error('queryVIPLoanFixedRateMarket() error:', error);
+    }
+}
+
+queryVIPLoanFixedRateMarket();